# docs/architecture/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: UnFranchise Marketing App API
  description: |
    API for the UnFranchise Marketing App - A corporate-controlled content sharing
    and engagement platform for UnFranchise Owners (UFOs).

    ## Authentication
    Most endpoints require Bearer token authentication via JWT.
    Include the access token in the Authorization header:
    ```
    Authorization: Bearer {access_token}
    ```

    ## Rate Limiting
    - Authentication endpoints: 5 requests per 15 minutes
    - Standard API endpoints: 100 requests per minute
    - Share endpoints: 500 requests per hour
    - Upload endpoints: 20 requests per hour

    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Pagination
    List endpoints support pagination via query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 20, max: 100)
    - `sort`: Sort field with optional `-` prefix for descending (e.g., `-createdAt`)

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.

  version: 1.0.0
  contact:
    name: API Support
    email: api-support@unfranchise.com
  license:
    name: Proprietary
servers:
  - url: https://api.unfranchise.com/v1
    description: Production server
  - url: https://staging-api.unfranchise.com/v1
...
paths:
...
  /auth/password/reset:
    post:
      tags: [Authentication]
      summary: Reset password
      description: Reset user password with token from email
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # User & Auth
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...